+++ src/validation.cpp
@@ -112,7 +112,7 @@ static constexpr int PRUNE_LOCK_BUFFER{10};
 
 GlobalMutex g_best_block_mutex;
 std::condition_variable g_best_block_cv;
-uint256 g_best_block;
+const CBlockIndex *g_best_block;
 
 BlockValidationOptions::BlockValidationOptions(const Config &config)
     : excessiveBlockSize(config.GetMaxBlockSize()), checkPoW(true),
@@ -2615,7 +2615,7 @@ void Chainstate::UpdateTip(const CBlockIndex *pindexNew) {
 
     {
         LOCK(g_best_block_mutex);
-        g_best_block = pindexNew->GetBlockHash();
+        g_best_block = pindexNew;
         g_best_block_cv.notify_all();
     }