+++ src/validation.h
@@ -115,7 +115,7 @@ enum class SynchronizationState { INIT_REINDEX, INIT_DOWNLOAD, POST_INIT };
 extern GlobalMutex g_best_block_mutex;
 extern std::condition_variable g_best_block_cv;
 /** Used to notify getblocktemplate RPC of new tips. */
-extern uint256 g_best_block;
+extern const CBlockIndex *g_best_block;
 
 /** Documentation for argument 'checklevel'. */
 extern const std::vector<std::string> CHECKLEVEL_DOC;